prototype/requirements.txt
streamlit==1.30.0
googlemaps==4.10.0
ortools==9.7.2996
pandas==2.1.1
matplotlib==3.8.0
firebase-admin==6.2.0
python-dotenv==1.0.0